chore(e2e/integrations/custom): bring the module-augmentation to the forefront
