+++ e2e/integrations/custom/src/index.mjs
@@ -1,0 +1,4 @@
+export * from "./example";
+export * from "./integration.mjs";
+// eslint-disable-next-line import-x/export
+export * from "./module-augmentation";
