+++ e2e/integrations/custom/src/module-augmentation/config.d.ts
@@ -1,0 +1,20 @@
+import type {
+  ExampleConfig,
+  ExampleResolvedConfig,
+} from "../example";
+
+declare module "@holypack/core"
+{
+  interface ConfigCustomProperties
+  {
+    example?: ExampleConfig | null;
+  }
+
+  interface ResolvedConfigCustomProperties
+  {
+    example: ExampleResolvedConfig;
+  }
+
+}
+
+export {};
