+++ e2e/integrations/custom/src/module-augmentation/context.d.ts
@@ -1,25 +1,13 @@
 import type {
-  ExampleConfig,
   ExampleContext,
-  ExampleResolvedConfig,
   ExampleResolvedContext,
-} from "./src/example";
+} from "../example";
 
 declare module "@holypack/core"
 {
-  interface ConfigCustomProperties
-  {
-    example?: ExampleConfig | null;
-  }
-
   interface ContextCustomProperties
   {
     example?: ExampleContext | null;
-  }
-
-  interface ResolvedConfigCustomProperties
-  {
-    example: ExampleResolvedConfig;
   }
 
   interface ResolvedContextCustomProperties
