+++ e2e/integrations/custom/src/module-augmentation/index.d.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line import-x/export
+export * from "./config";
+// eslint-disable-next-line import-x/export
+export * from "./context";
