+++ e2e/integrations/custom/tsconfig.scope.src.json
@@ -1,6 +1,3 @@
 {
-  "extends": "@holypack/integration-typescript/presets/tsconfig.scope.src.json",
-  "include": [
-    "./holypack.d.ts"
-  ]
+  "extends": "@holypack/integration-typescript/presets/tsconfig.scope.src.json"
 }
