+++ e2e/integrations/custom/types/holypack.d.ts
@@ -1,0 +1,1 @@
+import "../src/module-augmentation";
